fix(ui): address PR #281 review feedback and doctor CI failure
Fix disabled InputAction, obsolete README paragraph, chart series
color wrap-around, missing tooltip live region, and FieldHint/FieldError
coexistence flagged in review. Split ChartSeries/seriesColor out of
chart-parts.tsx and disable prefer-dynamic-import on recharts with
rationale, unblocking the react-doctor CI gate.
