+++ packages/ui/src/components/chart/bars-chart.tsx
@@ -1,3 +1,8 @@
+// Not loaded through next/dynamic, deliberately: this package has no
+// framework opinion, and recharts is already the only thing that pulls
+// these primitives in -- deferring the import buys nothing a consumer's own
+// code-splitting doesn't already give it.
+// react-doctor-disable-next-line react-doctor/prefer-dynamic-import
 import {
   Bar,
   BarChart,
@@ -7,14 +12,8 @@ import {
   XAxis,
   YAxis,
 } from 'recharts';
-import {
-  ChartLegend,
-  type ChartSeries,
-  ChartTooltip,
-  seriesColor,
-  XTick,
-  YTick,
-} from './chart-parts';
+import { ChartLegend, ChartTooltip, XTick, YTick } from './chart-parts';
+import { type ChartSeries, seriesColor } from './chart-series';
 
 /**
  * Bucketed bars over time: the events-by-severity chart, a log volume.
